moved functions that have a html and api context to their own files.
